+++ CHANGES.md
@@ -4,6 +4,7 @@ The release versions are PyPi releases.
 ## Version 3.3 (as yet unreleased)
 
 #### New Features
+  * Added fake `os.path.samefile` implementation ([#193](../../issues/193))
 
 #### Infrastructure